refactor: class, 함수 분리 및 export 구문 수정
